Fixes from PR comments.
