+++ src/main/resources/application.conf
@@ -1,14 +1,9 @@
-client {
-  client-exception = "Client Exception"
-  client-exception = ${?CLIENT_EXCEPTION}
-}
-
 auth {
   auth-header = "Authorization"
   auth-header = ${?AUTH_HEADER}
   token-type = "token"
   token-type = ${?TOKEN_TYPE}
-  token = "5cce42b5368aa7404972bb89ab3770f7f8035219"
+  token = "63f9d314732b1f36e67c139c2bd45c29e9f921ad"
   token = ${?TOKEN}
 }
 
@@ -30,7 +25,7 @@ repos {
   repos-filename = ${?REPOS_FILENAME}
   repos-header-line = "id,name,owner login,owner id,languages"
   repos-header-line = ${?REPOS_HEADER_LINE}
-  max-requests = 200
+  max-requests = 1000
   max-requests = ${?REPOS_MAX_REQUESTS}
 }
