+++ src/main/scala/com/app/sourcing/Failure.scala
@@ -1,0 +1,9 @@
+package com.app.sourcing
+
+trait Failure {
+  def exception: Throwable
+}
+
+object Failure {
+  final case class ClientException(msg: String, exception: Throwable) extends Throwable with Failure
+}
