+++ src/main/scala/com/app/sourcing/entity/GitHubFullUser.scala
@@ -1,0 +1,7 @@
+package com.app.sourcing.entity
+
+final case class GitHubFullUser(login: String, id: Long, name: String, location: String) {
+  override def toString: String = {
+    login + "," + id + "," + name + "," + location
+  }
+}
