+++ src/main/scala/com/app/sourcing/entity/GitHubRepo.scala
@@ -1,0 +1,3 @@
+package com.app.sourcing.entity
+
+final case class GitHubRepo(id: Long, languages_url: String)
