+++ src/main/scala/com/app/sourcing/entity/GitHubRepoOwner.scala
@@ -1,0 +1,3 @@
+package com.app.sourcing.entity
+
+final case class GitHubRepoOwner(login: String, id: Long)
