+++ src/main/scala/com/app/sourcing/entity/GitHubUser.scala
@@ -1,0 +1,3 @@
+package com.app.sourcing.entity
+
+final case class GitHubUser(login: String)
