+++ src/main/scala/com/app/sourcing/service/conf/AuthConf.scala
@@ -1,0 +1,13 @@
+package com.app.sourcing.service.conf
+
+trait AuthConf extends ConfigLoad {
+  def authHeader: String
+  def tokenType: String
+  def token: String
+}
+
+object AuthConf extends AuthConf {
+  def authHeader:      String = config.getString("auth.auth-header")
+  def tokenType:       String = config.getString("auth.token-type")
+  def token:           String = config.getString("auth.token")
+}
