+++ src/main/scala/com/app/sourcing/service/conf/ConfigLoad.scala
@@ -1,0 +1,7 @@
+package com.app.sourcing.service.conf
+
+import com.typesafe.config.{Config, ConfigFactory}
+
+trait ConfigLoad {
+  def config: Config = ConfigFactory.load()
+}
